Refactoring for 6.1 upgrade
Changes:
- Remove 'Rerun preupgrade check' button from the job
- Use :multiple_actions_provider for hosts bulk actions
