+++ app/helpers/concerns/foreman_leapp/remote_execution_helper_extension.rb
@@ -2,14 +2,11 @@
 
 module ForemanLeapp
   module RemoteExecutionHelperExtension
-    def job_invocation_task_buttons(task)
-      return super unless ::Helpers::JobHelper.correct_feature?(@job_invocation, 'leapp_remediation_plan')
-
-      super.insert(2, link_to(_('Rerun preupgrade check'),
-                              new_job_invocation_path(host_ids: @resource_base.map(&:id), feature: 'leapp_preupgrade'),
-                              class: 'btn btn-default',
-                              title: _('Run Leapp Preupgrade check again'),
-                              method: :get))
+    def rex_host_features(*args)
+      super + [link_to(_('Preupgrade check with Leapp'),
+                       new_job_invocation_path(host_ids: [args.first.id], feature: 'leapp_preupgrade')),
+               link_to(_('Upgrade with Leapp'),
+                       new_job_invocation_path(host_ids: [args.first.id], feature: 'leapp_upgrade'))]
     end
   end
 end
